[UI/#15] Bar 컴포저블 파일화
